pyproject.toml
[project]
name = "neural-vibe"
version = "0.1.0"
description = "Music discovery via brain response similarity using TRIBE v2"
requires-python = ">=3.12"
dependencies = [
    "tribev2 @ git+https://github.com/facebookresearch/tribev2.git",
    "faiss-cpu>=1.9.0",
    "click>=8.0",
    "rich>=13.0",
    "numpy>=2.0",
    "mutagen>=1.47",
]

[project.optional-dependencies]
finetune = [
    "openneuro-py>=2024.1",
    "nibabel>=5.0",
    "lightning>=2.0",
]
dev = [
    "pytest",
    "ipython",
]

[project.scripts]
neural-vibe = "neural_vibe.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/neural_vibe"]
